requirements.txt
notebook~=7.0.6
pandas~=2.1.4
rasterio~=1.3.9
requests~=2.31.0
python-dotenv~=1.0.0